requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.4
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
pycodestyle==2.6.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
